+++ src/app/error/error/error-routing.module.ts
@@ -1,0 +1,18 @@
+import { NgModule } from '@angular/core';
+import { RouterModule, Routes } from '@angular/router';
+import { ForbiddenComponent } from '../forbidden/forbidden.component';
+import { ErrorComponent } from './error.component';
+
+const routes: Routes = [
+  {
+    path: '',
+    component: ErrorComponent,
+    children: [{ path: 'forbidden', component: ForbiddenComponent }],
+  },
+];
+
+@NgModule({
+  imports: [RouterModule.forChild(routes)],
+  exports: [RouterModule],
+})
+export class ErrorRoutingModule {}