+++ src/app/error/forbidden/forbidden.component.ts
@@ -1,0 +1,10 @@
+import { Component, OnInit, ChangeDetectionStrategy } from '@angular/core';
+
+@Component({
+  templateUrl: './forbidden.component.html',
+  styleUrls: ['./forbidden.component.scss'],
+  changeDetection: ChangeDetectionStrategy.OnPush,
+})
+export class ForbiddenComponent {
+  constructor() {}
+}