+++ src/app/orders/orders.module.ts
@@ -1,8 +1,11 @@
 import { NgModule } from '@angular/core';
 import { BrowserModule } from '@angular/platform-browser';
+import { OrderComponent } from './order/order.component';
 
 @NgModule({
-  declarations: [],
+  declarations: [
+    OrderComponent
+  ],
   imports: [BrowserModule],
   exports: [],
   providers: [],