+++ src/app/products/components/product-card/product-card.component.ts
@@ -1,0 +1,23 @@
+import {
+  ChangeDetectionStrategy,
+  Component,
+  EventEmitter,
+  Input,
+  Output,
+} from '@angular/core';
+import { ProductModel } from '../../models/product.model';
+
+@Component({
+  selector: 'app-product-card',
+  templateUrl: './product-card.component.html',
+  styleUrls: ['./product-card.component.scss'],
+  changeDetection: ChangeDetectionStrategy.OnPush,
+})
+export class ProductCardComponent {
+  @Input() product!: ProductModel;
+  @Output() add = new EventEmitter<ProductModel>();
+
+  onAddToCart(): void {
+    this.add.emit(this.product);
+  }
+}