+++ src/app/products/products.module.ts
@@ -3,19 +3,13 @@ import { BrowserModule } from '@angular/platform-browser';
 import { SharedModule } from '../shared/shared.module';
 import { ProductListComponent } from './components/product-list/product-list.component';
 import { ProductComponent } from './components/product/product.component';
+import { ProductCardComponent } from './components/product-card/product-card.component';
+import { RouterModule } from '@angular/router';
 
 @NgModule({
-  declarations: [
-    ProductComponent,
-    ProductListComponent,
-  ],
-  imports: [
-    BrowserModule,
-    SharedModule
-  ],
-  exports:[
-    ProductListComponent,
-  ],
+  declarations: [ProductComponent, ProductListComponent, ProductCardComponent],
+  imports: [BrowserModule, SharedModule, RouterModule],
+  exports: [ProductListComponent],
   providers: [],
 })
-export class ProductsModule { }
+export class ProductsModule {}